pyproject.toml
[project]
name = "buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "fastembed>=0.7.4",
    "google-genai>=1.64.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "pandas>=2.3.3",
    "pymongo[srv]==3.12",
    "qdrant-client>=1.16.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
